Rename "flyweight-oracle-node" project to "triggers-oracle-node". Add new oracle node to verify on-chain deposits to the flyweight smart contract
